🎨 rewrote example
