+++ rust/tokens_transfer/Cargo.toml
@@ -1,22 +1,5 @@
-[package]
-name = "tokens_transfer"
-version = "0.1.0"
-edition = "2018"
-
-## See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[lib]
-crate-type = ["cdylib"]
-
-[build-dependencies]
-candid = "0.7.4"
-
-[dependencies]
-candid = "0.7.4"
-ic-cdk = "0.3.3"
-ic-cdk-macros = "0.3.3"
-ic-ledger-types = "0.1.0"
-serde = "1.0.126"
-serde_derive = "1.0.126"
-
-[dev-dependencies]
-tempfile = "3.2.0"
+[workspace]
+members = [
+    "src/tokens_transfer_backend"
+]
+resolver = "2"
