--- rust/tokens_transfer/src/lib.rs
@@ -1,66 +1,0 @@
-use std::cell::RefCell;
-use std::hash::Hash;
-use candid::{candid_method, CandidType, Principal};
-
-use ic_cdk_macros::*;
-use ic_ledger_types::{AccountIdentifier, BlockIndex, DEFAULT_SUBACCOUNT, MAINNET_LEDGER_CANISTER_ID, Memo, Subaccount, Tokens};
-use serde::{Deserialize, Serialize};
-
-#[derive(CandidType, Serialize, Deserialize, Clone, Debug, Hash, PartialEq)]
-pub struct Conf {
-    ledger_canister_id: Principal,
-    // The subaccount of the account identifier that will be used to withdraw tokens and send them
-    // to another account identifier. If set to None then the default subaccount will be used.
-    // See the [Ledger doc](https://internetcomputer.org/docs/current/developer-docs/integrations/ledger/#accounts).
-    subaccount: Option<Subaccount>,
-    transaction_fee: Tokens
-}
-
-impl Default for Conf {
-    fn default() -> Self {
-        Conf {
-            ledger_canister_id: MAINNET_LEDGER_CANISTER_ID,
-            subaccount: None,
-            transaction_fee: Tokens::from_e8s(10_000),
-        }
-    }
-}
-
-thread_local! {
-    static CONF: RefCell<Conf> = RefCell::new(Conf::default());
-}
-
-#[init]
-#[candid_method(init)]
-fn init(conf: Conf) {
-    CONF.with(|c| c.replace(conf));
-}
-
-#[derive(CandidType, Serialize, Deserialize, Clone, Debug, Hash)]
-pub struct TransferArgs {
-    amount: Tokens,
-    to_principal: Principal,
-    to_subaccount: Option<Subaccount>,
-}
-
-#[update]
-#[candid_method(update)]
-async fn transfer(args: TransferArgs) -> Result<BlockIndex, String> {
-    ic_cdk::println!("Transferring {} tokens to principal {} subaccount {:?}", &args.amount, &args.to_principal, &args.to_subaccount);
-    let ledger_canister_id = CONF.with(|conf| conf.borrow().ledger_canister_id);
-    let to_subaccount = args.to_subaccount.unwrap_or(DEFAULT_SUBACCOUNT);
-    let transfer_args = CONF.with(|conf| {
-        let conf = conf.borrow();
-        ic_ledger_types::TransferArgs {
-            memo: Memo(0),
-            amount: args.amount,
-            fee: conf.transaction_fee,
-            from_subaccount: conf.subaccount,
-            to: AccountIdentifier::new(&args.to_principal, &to_subaccount),
-            created_at_time: None,
-        }
-    });
-    ic_ledger_types::transfer(ledger_canister_id, transfer_args).await
-        .map_err(|e| format!("failed to call ledger: {:?}", e))?
-        .map_err(|e| format!("ledger transfer error {:?}", e))
-}
