+++ rust/tokens_transfer/src/tokens_transfer_backend/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "tokens_transfer_backend"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+candid = "0.10.4"
+ic-cdk = "0.12.1"
+ic-cdk-macros = "0.8.4"
+ic-ledger-types = "0.9.0"
+serde = "1.0.197"
+serde_derive = "1.0.197"
